+++ BattleShip/BattleShip/Debug/BattleShip.log
@@ -1,7 +1,9 @@
 ﻿  Main.cpp
-c:\users\rlawn\desktop\wowproject\battleship\battleship\main.cpp(79): warning C4996: 'scanf': This function or variable may be unsafe. Consider using scanf_s instead. To disable deprecation, use _CRT_SECURE_NO_WARNINGS. See online help for details.
+c:\users\rlawn\desktop\wowproject\battleship\battleship\main.cpp : warning C4819: 현재 코드 페이지(949)에서 표시할 수 없는 문자가 파일에 들어 있습니다. 데이터가 손실되지 않게 하려면 해당 파일을 유니코드 형식으로 저장하십시오.
+c:\users\rlawn\desktop\wowproject\battleship\battleship\main.cpp(84): warning C4996: 'scanf': This function or variable may be unsafe. Consider using scanf_s instead. To disable deprecation, use _CRT_SECURE_NO_WARNINGS. See online help for details.
   c:\program files (x86)\windows kits\10\include\10.0.10240.0\ucrt\stdio.h(1270): note: 'scanf' 선언을 참조하십시오.
-c:\users\rlawn\desktop\wowproject\battleship\battleship\main.cpp(157): warning C4996: 'scanf': This function or variable may be unsafe. Consider using scanf_s instead. To disable deprecation, use _CRT_SECURE_NO_WARNINGS. See online help for details.
+c:\users\rlawn\desktop\wowproject\battleship\battleship\main.cpp(162): warning C4996: 'scanf': This function or variable may be unsafe. Consider using scanf_s instead. To disable deprecation, use _CRT_SECURE_NO_WARNINGS. See online help for details.
   c:\program files (x86)\windows kits\10\include\10.0.10240.0\ucrt\stdio.h(1270): note: 'scanf' 선언을 참조하십시오.
+c:\users\rlawn\desktop\wowproject\battleship\battleship\main.cpp(1191): warning C4819: 현재 코드 페이지(949)에서 표시할 수 없는 문자가 파일에 들어 있습니다. 데이터가 손실되지 않게 하려면 해당 파일을 유니코드 형식으로 저장하십시오.
   BattleShip.vcxproj -> C:\Users\rlawn\Desktop\WOWProject\BattleShip\Debug\BattleShip.exe
   BattleShip.vcxproj -> C:\Users\rlawn\Desktop\WOWProject\BattleShip\Debug\BattleShip.pdb (Full PDB)
