+++ BattleShipServer/bin/app.js
@@ -1,7 +1,7 @@
 var http = require('http');
 var express = require('express');
 var app = express();
-var port = 3000;
+var port = 80;
 var httpServer = http.createServer(app).listen(port, function (req, res) {
 });
