+++ Test/stack.c
@@ -1,75 +1,29 @@
-#include <stdio.h>
-#define SIZE 5
-int top=0;
-int s[SIZE];
 
+#include "stdafx.h"
+#include "stdlib.h"
+#define N 100
 
-void push(int d)
+char S[N] = {};
+int top = -1;
+
+void push(char c);
+char pop();
+
+int main()
 {
-	if(top >= SIZE) {printf("overflow\n");
-	}
-	
-	else {
-		s[top] = d;
-		top++;
-	}
-	
+
+
+	system("pause");
+	return 0;
 }
 
-void pop()
+void push(char c)
 {
-	if(top == 0) {printf("empty\n"); }		
-	else {
-		printf("pop된 값은 ? %d\n", s[top-1]);
-		s[top-1]=0;
-		top--;
-		
-	}
+	top++;
+	S[top] = c;
 }
 
-
-void dis()
+char pop()
 {
-	int i;
-	for(i=0;i<SIZE;i++)
-	{
-		printf("%d\t",s[i]);
-	}
-	printf("\n\n");
-}
-
-
-void main()
-{
-	int menu;
-	int data;
-	
-
-
-	while(menu <= 3)
-	{	printf("메뉴 선택 1. 푸시    2. 팝    3. 종료\n");
-		scanf("%d", &menu);
-		
-		switch (menu)
-		{
-		case 1:
-			printf("스택에 푸시할 값을 입력하세요");
-			scanf("%d",&data);
-			push(data);
-			break;
-		case 2:
-			pop();
-			break;
-		case 3:
-			menu = 5;
-			break;
-		default :
-			printf("잘못된 메뉴 입력");
-			break;
-
-		}
-		dis();
-	}
-	
-
+	return top--;
 }
